# swaggerdoc.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Medten Product API'
  description: 'Interact with the Medten product database using this API.'
servers:
  - url: https://hm5fehgga2.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API Test Environment
  - url: https://w5d9g2jvp2.execute-api.us-west-1.amazonaws.com/Prod/api/v1/
    description: API Production Environment
  - url: http://localhost:3000/api/v1/
    description: Developer Local Test Environments
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/gaumeister/medten-product-api/1.0.0
paths: 
...
  /brands:
    summary: Brands for products.
    parameters:
    - name: id
      in: query
      description: a specific brand id
      required: false
      schema:
        type: integer
    - name: name_en
      in: query
      description: name of the brand (English)
      required: false
      schema:
        type: string
    - name: name_zh
      in: query
      description: name of the brand (Chinese)
      required: false
      schema:
        type: string
    get:
      summary: searches for brands
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  brands:
                    type: array
                    items:
                      $ref: '#/components/schemas/Brand'

        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /brands/${id}:
    parameters:
    - name: id
      in: path
      description: the brand id
      required: true
      schema:
        type: integer
    get:
      summary: gets a brand by id
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Brand'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    put:
      summary: updates a brand
      tags:
        - general
      requestBody:
        description: >
          The brand to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Brand'
      responses:
        200:
          description: returns the brand 
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Brand'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /categories:
    summary: Categories of medical device products.
    parameters:
    - name: id
      in: query
      description: a specific category id
      required: false
      schema:
        type: integer
    - name: name_en
      in: query
      description: name of the category (English)
      required: false
      schema:
        type: string
    - name: featured
      in: query
      description: whether the category is featured in the Catalog (1) or not (0)
      required: false
      schema:
        type: integer
    get:
      summary: searches for categories
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_views:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryView'

        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /custom_attributes:
    summary: Custom attributes that may be assigned to products.
    parameters:
    - name: id
      in: query
      description: a specific custom attribute id
      required: false
      schema:
        type: integer
    - name: name_en
      in: query
      description: name of the custom attribute (English)
      required: false
      schema:
        type: string
    - name: name_zh
      in: query
      description: name of the custom attribute (Chinese)
      required: false
      schema:
        type: string
    - name: category_id
      in: query
      description: a specific category_id
      required: false
      schema:
        type: integer
    get:
      summary: searches for custom attributes
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  custom_attributes:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomAttribute'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /equipment_models:
    summary: Models used across all equipment.
    description: An equipment is any entity having a model number that could potentially be assigned to a product.
    parameters:
    - name: brand_id
      in: query
      description: the brand id to narrow the search
      required: false
      schema:
        type: integer
    - name: model_search
      in: query
      description: the search string to perform a wildcard search across all models 
      required: false
      schema:
        type: string
    get:
      summary: searches for equipment models
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  equipment_models:
                    type: array
                    items:
                      type: object
                      properties:
                        model:
                          type: string
                        brand_id:
                          type: integer
                        brand_en:
                          type: string
                        brand_zh:
                          type: string

        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /filter_option_views:
    summary: Filtering options available across products to facilitate precise searches.
    parameters:
    - name: category_id
      in: query
      description: a specific category id
      required: false
      schema:
        type: integer
    - name: category_en
      in: query
      description: name of the category (English)
      required: false
      schema:
        type: string
    - name: category_zh
      in: query
      description: name of the category (Chinese)
      required: false
      schema:
        type: string
    - name: filter_id
      in: query
      description: the id of the filter
      required: false
      schema:
        type: integer
    - name: visible_in_catalog
      in: query
      description: flag indicating whether the filter is used (visible) in the Product Catalog application for filtering
      required: false
      schema:
        type: boolean 
    - name: filter_option_id
      in: query
      description: the individual filter option id
      required: false
      schema:
        type: integer
    - name: filter_en
      in: query
      description: name of the filter (English)
      required: false
      schema:
        type: string
    - name: filter_zh
      in: query
      description: name of the filter (Chinese)
      required: false
      schema:
        type: string
    - name: option_en
      in: query
      description: name of the option (English)
      required: false
      schema:
        type: string
    - name: option_zh
      in: query
      description: name of the option (Chinese)
      required: false
      schema:
        type: string
    get:
      summary: search for filter options based on criteria other than product id
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/ProductFilterOptionView'
  /marketing_regions:
    summary: Marketing regions.
    parameters:
    - name: id
      in: query
      description: a specific marketing region id
      required: false
      schema:
        type: integer
    - name: name_en
      in: query
      description: name of the marketing region (English)
      required: false
      schema:
        type: string
    get:
      summary: searches for marketing regions
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                   marketing_regions:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketingRegion'

        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /suppliers:
    summary: Suppliers of equipment.
    parameters:
    - name: id
      in: query
      description: a specific supplier id
      required: false
      schema:
        type: integer
    - name: name_en
      in: query
      description: name of the supplier (Englisth)
      required: false
      schema:
        type: string
    - name: name_zh
      in: query
      description: name of the supplier (Chinese)
      required: false
      schema:
        type: string
    get:
      summary: searches for suppliers
      tags:
        - general
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  suppliers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Supplier'

        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
  /groups:
    summary: Represents a product group
    post:
      summary: creates a group
      tags:
        - groups
      requestBody:
        description:  The group to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        200:
          description: returns the group 
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Group'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: A human-readable message describing the error.
        error:
          type: string
          description: The error details. Typically error details may contain sensitive information not intended to be shown directly in the user interface.
    Auditable:
      type: object
      properties:
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        version:
          type: integer
          readOnly: true
    Brand:
      type: object
      description: An original equipment manufacturer (OEM) or brand.
      allOf:
        - type: object
          properties:
            id:
              type: integer
              readOnly: true
            name_en:
              type: string
              maxLength: 255
            name_zh:
              type: string
              maxLength: 255
            parent_id:
              type: integer
            is_oem:
              type: boolean
        - $ref: '#/components/schemas/Auditable'
    Category:
      type: object
      description: Common grouping used to categorize families. Categories define certain defaults and constraints across families.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          name_en:
            description: English name.
            type: string
            maxLength: 255
          name_zh:
            description: Chinese name.
            type: string
            maxLength: 255
          parent_id:
            description: ID of containing parent category
            type: integer
          product_name_formula_id:
            description: Identifies a stored **Formula** that can be used to auto-generate product names (title)
            type: string
            maxLength: 255
          product_description_formula_id:
            description: Identifies a stored **Formula** that can be used to auto-generate product descriptions.
            type: string
            maxLength: 255
          family_name_formula_id:
            description: Identifies a stored **Formula** that can be used to auto-generate family names (title).
            type: string
            maxLength: 255
          valid_image_types:
            type: string
            maxLength: 255
          featured:
            description: Identifies whether the category should be featured in the Catalog.
            type: boolean
          image_url:
            description: Specifies a category image to graphically depict the category in the Catalog.
            type: string
            maxLength: 255
      - $ref: '#/components/schemas/Auditable'
    CategoryView:
      type: object
      description: Extended view of the Category. It pulls in the naming-formulas associated with the category.
      allOf:
      - type: object
        properties:
          
          product_name_formula:
            description: javascript formula for generating product names
            type: string
            maxLength: 4000
          product_description_formula:
            description: javascript formula for generating product descriptions
            type: string
            maxLength: 4000
          family_name_formula:
            description: javascript formula for generating family names (title).
            type: string
            maxLength: 4000
      - $ref: '#/components/schemas/Category'
      - $ref: '#/components/schemas/Auditable'
...
    CustomAttribute:
      type: object
      description: A custom attribute that can be assigned to a product
      allOf:
        - type: object
          properties:
            id:
              type: integer
              readOnly: true
            name_en:
              type: string
              maxLength: 255
            name_zh:
              type: string
              maxLength: 255
            category_id:
              type: integer
        - $ref: '#/components/schemas/Auditable'
...
    Group:
      type: object
      description: The group used to group together families.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          group_code:
            description: Code for the group
            type: string
            maxLength: 255
      - $ref: '#/components/schemas/Auditable'
...
    MarketingRegion:
      type: object
      allOf:
        - type: object
          properties:
            name_en:
              type: string
            id:
              type: integer
              readOnly: true
        - $ref: '#/components/schemas/Auditable'
...
    ProductFilterOptionView:
      type: object
      description: The filter option reference for a product, including extended properties.
      allOf:
      - type: object
        properties:
          id:
            type: integer
          product_id:
            type: integer
          filter_id:
            description: reference to the filter associated with the filter option
            type: integer
          filter_en:
            description: The name of the filter (English)
            type: string
          filter_zh:
            description: The name of the filter (Chinese)
            type: string
          filter_option_id:
            type: integer
          option_en:
            description: The name of the option (English)
            type: string
          option_zh:
            description: The name of the option (Chinese)
            type: string
          option_us:
            description: The name of the option (US English). Typically only used if imperial units are needed for an option.
            type: string
          priority_order:
            description: the display order/priority of the option
            type: integer
      - $ref: '#/components/schemas/Auditable'
...
    Supplier:
      type: object
      allOf:
        - type: object
          properties:
            name_en:
              type: string
            name_zh:
              type: string
            id:
              type: integer
              readOnly: true
        - $ref: '#/components/schemas/Auditable'
  responses:
    400ClientError:
      description: client error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'
...
    500ServerError:
      description: server error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'